replaces routing with auth util, setting mobileauth in session
